split S1 in two.
