+++ msresist/figures/figureS2.py
@@ -1,22 +1,18 @@
 """
-This creates Figure 2.
+This creates Supplemental Figure 2: Cell migration and island
 """
 
-import pandas as pd
+import numpy as np
 import seaborn as sns
 from .common import subplotLabel, getSetup
-from ..pre_processing import preprocessing
-from .figure1 import plot_IdSites, plot_AllSites, plotPCA_scoresORloadings
-from ..motifs import MapMotifs
-
-
-all_lines = ["WT", "KO", "KD", "KI", "Y634F", "Y643F", "Y698F", "Y726F", "Y750F", "Y821F"]
+from .figure1 import IndividualTimeCourses, import_phenotype_data, barplot_UtErlAF154
+from ..distances import BarPlotRipleysK, PlotRipleysK
 
 
 def makeFigure():
     """Get a list of the axis objects and create a figure"""
     # Get list of axis objects
-    ax, f = getSetup((12, 11), (2, 3), multz={0: 1})
+    ax, f = getSetup((15, 10), (4, 6), multz={0:1, 12:1})
 
     # Add subplot labels
     subplotLabel(ax)
@@ -24,24 +20,26 @@ def makeFigure():
     # Set plotting format
     sns.set(style="whitegrid", font_scale=1.2, color_codes=True, palette="colorblind", rc={"grid.linestyle": "dotted", "axes.linewidth": 0.6})
 
-    # Heatmap Signaling
-    ax[0].axis("off")
+    # Read in migration data
+    sw = import_phenotype_data(phenotype="Migration")
 
-    # Read in Mass Spec data
-    X = preprocessing(Axlmuts_ErlAF154=True, Vfilter=True, FCfilter=True, log2T=True, mc_col=True)
+    # Labels
+    lines = ["WT", "KO", "KI", "KD", "Y634F", "Y643F", "Y698F", "Y726F", "Y750F", "Y821F"]
+    tr1 = ["-UT", "-E", "-A/E"]
+    tr2 = ["Untreated", "Erlotinib", "Erl + AF154"]
+    t1 = ["UT", "AF", "-E", "A/E"]
+    t2 = ["Untreated", "AF154", "Erlotinib", "Erl + AF154"]
+    colors = ["white", "windows blue", "scarlet"]
+    mutants = ['PC9', 'KO', 'KIN', 'KD', 'M4', 'M5', 'M7', 'M10', 'M11', 'M15']
+    itp = 24
 
-    # PCA
-    data = X.set_index(["Gene"]).select_dtypes(include=float)
-    data.columns = all_lines
-    plotPCA_scoresORloadings(ax[1], data.reset_index(), 3, ["Gene"], "Signaling", plot="loadings")
+    # Bar plots
+    barplot_UtErlAF154(ax[0], lines, sw, 14, tr1, tr2, "fold-change RWD", "Cell Migration (t=14h)", TreatmentFC="-E", colors=colors, TimePointFC=itp)
+    BarPlotRipleysK(ax[11], '48hrs', mutants, lines, ['ut', 'e', 'ae'], tr2, 6, np.linspace(1.5, 14.67, 1), colors, TreatmentFC="Erlotinib", ylabel="fold-change K estimate")
 
-    # Specific p-sites
-    erk = {"MAPK1": "Y187-p", "MAPK3": "Y204-p"}
-    erk_rn = ["ERK2", "ERK1"]
-
-    plot_AllSites(ax[2], X.copy(), "AXL", "AXL", ylim=[-3, 2.5])
-    ax[3].legend(loc='upper left', prop={'size': 8})
-    plot_AllSites(ax[3], X.copy(), "EGFR", "EGFR", ylim=[-3, 2.5])
-    plot_IdSites(ax[4], X.copy(), erk, "ERK1/2", rn=erk_rn, ylim=[-3, 2.5])
+    # Time courses
+    for i, line in enumerate(lines):
+        IndividualTimeCourses(sw, 24, lines, t1, t2, "RWD %", plot=line, ax_=ax[i+1])
+        PlotRipleysK('48hrs', mutants[i], ['ut', 'e', 'ae'], 6, ax=ax[i+12], title=line)
 
     return f
